move table fetching to app
