+++ frontend/src/components/plot.tsx
@@ -1,30 +1,13 @@
 import { BarChart, XAxis, YAxis, Tooltip, Bar, Label } from "recharts"
 import { useTheme } from "@material-ui/core"
-import { useAsync } from "react-async-hook"
-import { ParticipantV, WithdrawnV } from "../lib/data"
-import { tableFetch, useTableData } from "../lib/table-data"
+import { Participant } from "../lib/data"
 import * as d3 from "d3-array"
 
 export default function Plots({
-  token,
-  withdrawnSetting,
+  participants,
 }: {
-  token?: string
-  withdrawnSetting: "yes" | "no" | "any"
+  participants: Participant[]
 }) {
-  const participantsFetch = useAsync(tableFetch, [
-    "participants",
-    ParticipantV,
-    token,
-  ])
-  const withdrawnFetch = useAsync(tableFetch, ["withdrawn", WithdrawnV, token])
-
-  const withdrawn = useTableData(withdrawnFetch.result, {})
-  const tableSettings = {
-    withdrawn: { setting: withdrawnSetting, ids: withdrawn.map((w) => w.pid) },
-  }
-  const participants = useTableData(participantsFetch.result, tableSettings)
-
   const genderCounts = d3.rollup(
     participants,
     (v) => v.length,
